// src/lib/conduit-client/models/SingleArticleResponse.ts
// tslint:disable
/**
 * Conduit API
 * Conduit API
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Article } from './'

/**
 * @export
 * @interface SingleArticleResponse
 */
export interface SingleArticleResponse {
  /**
   * @type {Article}
   * @memberof SingleArticleResponse
   */
  article: Article
}