// src/lib/conduit-client/models/UpdateArticle.ts
// tslint:disable
/**
 * Conduit API
 * Conduit API
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface UpdateArticle
 */
export interface UpdateArticle {
  /**
   * @type {string}
   * @memberof UpdateArticle
   */
  title?: string
  /**
   * @type {string}
   * @memberof UpdateArticle
   */
  description?: string
  /**
   * @type {string}
   * @memberof UpdateArticle
   */
  body?: string
  /**
   * @type {Array<string>}
   * @memberof UpdateArticle
   */
  tagList?: Array<string>
}